Refactor option greeks feature
