+++ nautilus_trader/common/actor.pyx
@@ -26,6 +26,7 @@ attempts to operate without a managing `Trader` instance.
 
 import asyncio
 from concurrent.futures import Executor
+from typing import Callable
 
 import cython
 
@@ -34,8 +35,10 @@ from nautilus_trader.common.config import ImportableActorConfig
 from nautilus_trader.common.executor import ActorExecutor
 from nautilus_trader.common.executor import TaskId
 from nautilus_trader.common.signal import generate_signal_class
-from nautilus_trader.model.greeks import GreeksData
-from nautilus_trader.model.greeks import PortfolioGreeks
+from nautilus_trader.model.data import Bar
+from nautilus_trader.model.data import DataType
+from nautilus_trader.model.enums import PriceType
+from nautilus_trader.model.identifiers import InstrumentId
 
 from cpython.datetime cimport datetime
 from libc.stdint cimport uint64_t
@@ -54,7 +57,6 @@ from nautilus_trader.core.message cimport Event
 from nautilus_trader.core.rust.common cimport ComponentState
 from nautilus_trader.core.rust.common cimport LogColor
 from nautilus_trader.core.rust.model cimport BookType
-from nautilus_trader.core.rust.model cimport PositionSide
 from nautilus_trader.core.rust.model cimport PriceType
 from nautilus_trader.core.uuid cimport UUID4
 from nautilus_trader.data.messages cimport DataRequest
@@ -71,10 +73,10 @@ from nautilus_trader.model.data cimport OrderBookDelta
 from nautilus_trader.model.data cimport OrderBookDeltas
 from nautilus_trader.model.data cimport QuoteTick
 from nautilus_trader.model.data cimport TradeTick
+from nautilus_trader.model.greeks cimport GreeksCalculator
 from nautilus_trader.model.identifiers cimport ClientId
 from nautilus_trader.model.identifiers cimport ComponentId
 from nautilus_trader.model.identifiers cimport InstrumentId
-from nautilus_trader.model.identifiers cimport StrategyId
 from nautilus_trader.model.identifiers cimport Venue
 from nautilus_trader.model.instruments.base cimport Instrument
 from nautilus_trader.model.instruments.synthetic cimport SyntheticInstrument
@@ -120,7 +122,6 @@ cdef class Actor(Component):
         self._warning_events: set[type] = set()
         self._pending_requests: dict[UUID4, Callable[[UUID4], None] | None] = {}
         self._pyo3_conversion_types = set()
-        self._future_greeks: dict[InstrumentId, list[GreeksData]] = {}
         self._signal_classes: dict[str, type] = {}
 
         # Indicators
@@ -138,6 +139,7 @@ cdef class Actor(Component):
         self.msgbus = None     # Initialized when registered
         self.cache = None      # Initialized when registered
         self.clock = None      # Initialized when registered
+        self.greeks = None     # Initialized when registered
         self.log = self._log
 
     def to_importable_config(self) -> ImportableActorConfig:
@@ -582,6 +584,8 @@ cdef class Actor(Component):
         self.cache = cache
         self.clock = self._clock
         self.log = self._log
+
+        self.greeks = GreeksCalculator(msgbus, cache, self.clock, self.log)
 
     cpdef void register_executor(
         self,
@@ -3374,87 +3378,3 @@ cdef class Actor(Component):
         if is_logging_initialized():
             self._log.info(f"{REQ}{SENT} {request}")
         self._msgbus.request(endpoint="DataEngine.request", request=request)
-
-# -- GREEKS ---------------------------------------------------------------------------------------
-
-    def instrument_greeks_data(self, InstrumentId instrument_id) -> GreeksData:
-        """
-        Retrieve the Greeks data for a given instrument.
-
-        This method handles both options and futures instruments. For options,
-        it retrieves the Greeks data from the cache. For futures, it creates
-        a GreeksData object based on the instrument's delta and multiplier.
-
-        Parameters
-        ----------
-        instrument_id : InstrumentId
-            The identifier of the instrument for which to retrieve Greeks data.
-
-        Returns
-        -------
-        GreeksData
-            The Greeks data for the specified instrument, including vol, price, delta, gamma, vega, theta.
-
-        """
-        from nautilus_trader.risk.greeks import greeks_key
-
-        # Option case, to avoid querying definition
-        if " " in instrument_id.symbol.value:
-            return GreeksData.from_bytes(self.cache.get(greeks_key(instrument_id)))
-
-        # Future case
-        if instrument_id not in self._future_greeks:
-            future_definition = self.cache.instrument(instrument_id)
-            self._future_greeks[instrument_id] = GreeksData.from_delta(instrument_id, int(future_definition.multiplier))
-
-        return self._future_greeks[instrument_id]
-
-    def portfolio_greeks(
-        self, str underlying = "",
-        Venue venue = None,
-        InstrumentId instrument_id = None,
-        StrategyId strategy_id = None,
-        PositionSide side = PositionSide.NO_POSITION_SIDE,
-    ) -> PortfolioGreeks:
-        """
-        Calculate the portfolio Greeks for a given set of positions.
-
-        This method aggregates the Greeks data for all open positions that match the specified criteria.
-
-        Parameters
-        ----------
-        underlying : str, optional
-            The underlying asset symbol to filter positions. If provided, only positions with instruments
-            starting with this symbol will be included. Default is an empty string (no filtering).
-        venue : Venue, optional
-            The venue to filter positions. If provided, only positions from this venue will be included.
-        instrument_id : InstrumentId, optional
-            The instrument ID to filter positions. If provided, only positions for this instrument will be included.
-        strategy_id : StrategyId, optional
-            The strategy ID to filter positions. If provided, only positions for this strategy will be included.
-        side : PositionSide, optional
-            The position side to filter. If provided, only positions with this side will be included.
-            Default is PositionSide.NO_POSITION_SIDE (no filtering).
-
-        Returns
-        -------
-        PortfolioGreeks
-            The aggregated Greeks data for the portfolio, including delta, gamma, vega, theta.
-
-        """
-        ts_event = self.clock.timestamp_ns()
-        portfolio_greeks = PortfolioGreeks(ts_event, ts_event)
-        open_positions = self.cache.positions_open(venue, instrument_id, strategy_id, side)
-
-        for position in open_positions:
-            position_instrument_id = position.instrument_id
-
-            if underlying != "" and not position_instrument_id.value.startswith(underlying):
-                continue
-
-            quantity = float(position.signed_qty)
-            instrument_greeks = self.instrument_greeks_data(position_instrument_id)
-            position_greeks = quantity * instrument_greeks
-            portfolio_greeks += position_greeks
-
-        return portfolio_greeks
