+++ nautilus_trader/model/greeks.pxd
@@ -1,0 +1,30 @@
+# -------------------------------------------------------------------------------------------------
+#  Copyright (C) 2015-2025 Nautech Systems Pty Ltd. All rights reserved.
+#  https://nautechsystems.io
+#
+#  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
+#  You may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+# -------------------------------------------------------------------------------------------------
+
+from nautilus_trader.cache.base cimport CacheFacade
+from nautilus_trader.common.component cimport Clock
+from nautilus_trader.common.component cimport Logger
+from nautilus_trader.common.component cimport MessageBus
+
+
+cdef class GreeksCalculator:
+    cdef readonly Clock clock
+    """The actors clock.\n\n:returns: `Clock`"""
+    cdef readonly Logger log
+    """The actors logger.\n\n:returns: `Logger`"""
+    cdef readonly MessageBus msgbus
+    """The message bus for the actor (if registered).\n\n:returns: `MessageBus` or ``None``"""
+    cdef readonly CacheFacade cache
+    """The read-only cache for the actor.\n\n:returns: `CacheFacade`"""
